Implementing tests
