+++ blockchain/src/blockchain/verify.rs
@@ -127,29 +127,31 @@ impl Blockchain {
             }
         }
 
-        // Verify the history root.
-        let real_history_root = self
-            .history_store
-            .get_history_tree_root(block.epoch_number(), txn_opt)
-            .ok_or_else(|| {
-                error!(
+        // Verify the history root when we have the full transaction history of the epoch.
+        if state.can_verify_history {
+            let real_history_root = self
+                .history_store
+                .get_history_tree_root(block.epoch_number(), txn_opt)
+                .ok_or_else(|| {
+                    error!(
+                        %block,
+                        epoch_number = block.epoch_number(),
+                        reason = "failed to fetch history tree root for epoch from store",
+                        "Rejecting block"
+                    );
+                    PushError::InvalidBlock(BlockError::InvalidHistoryRoot)
+                })?;
+
+            if *block.history_root() != real_history_root {
+                warn!(
                     %block,
-                    epoch_number = block.epoch_number(),
-                    reason = "failed to fetch history tree root for epoch from store",
+                    block_root = %block.history_root(),
+                    history_root = %real_history_root,
+                    reason = "History root doesn't match real history root",
                     "Rejecting block"
                 );
-                PushError::InvalidBlock(BlockError::InvalidHistoryRoot)
-            })?;
-
-        if *block.history_root() != real_history_root {
-            warn!(
-                %block,
-                block_root = %block.history_root(),
-                history_root = %real_history_root,
-                reason = "History root doesn't match real history root",
-                "Rejecting block"
-            );
-            return Err(PushError::InvalidBlock(BlockError::InvalidHistoryRoot));
+                return Err(PushError::InvalidBlock(BlockError::InvalidHistoryRoot));
+            }
         }
 
         // For macro blocks we have additional checks. We simply construct what the body should be
