+++ consensus/tests/sync_utils.rs
@@ -1,0 +1,269 @@
+use std::path::PathBuf;
+use std::sync::Arc;
+
+use futures::{future, StreamExt};
+use nimiq_blockchain_proxy::BlockchainProxy;
+use nimiq_bls::cache::PublicKeyCache;
+use nimiq_consensus::sync::live::block_queue::{BlockHeaderTopic, BlockTopic};
+use nimiq_consensus::sync::syncer::MacroSyncReturn;
+use nimiq_consensus::sync::syncer_proxy::SyncerProxy;
+use nimiq_light_blockchain::LightBlockchain;
+use nimiq_test_utils::node::TESTING_BLS_CACHE_MAX_CAPACITY;
+use nimiq_test_utils::zkp_test_data::{zkp_test_exe, KEYS_PATH};
+use nimiq_zkp_component::ZKPComponent;
+use parking_lot::{Mutex, RwLock};
+
+use nimiq_block_production::BlockProducer;
+use nimiq_blockchain::{Blockchain, BlockchainConfig};
+use nimiq_blockchain_interface::{AbstractBlockchain, BlockchainEvent};
+use nimiq_consensus::consensus::Consensus;
+use nimiq_database::volatile::VolatileEnvironment;
+use nimiq_genesis::NetworkId;
+use nimiq_network_interface::network::Network as NetworkInterface;
+use nimiq_network_libp2p::Network;
+use nimiq_network_mock::MockHub;
+use nimiq_test_utils::{
+    blockchain::{produce_macro_blocks_with_txns, signing_key, voting_key},
+    test_network::TestNetwork,
+};
+use nimiq_utils::time::OffsetTime;
+use tokio::spawn;
+
+#[allow(dead_code)]
+pub enum SyncMode {
+    History,
+    Full,
+    Light,
+}
+
+async fn syncer(
+    sync_mode: &SyncMode,
+    network: &Arc<Network>,
+    blockchain: &BlockchainProxy,
+    zkp_prover: &ZKPComponent<Network>,
+) -> SyncerProxy<Network> {
+    match sync_mode {
+        SyncMode::History => {
+            SyncerProxy::new_history(
+                blockchain.clone(),
+                Arc::clone(&network),
+                Arc::new(Mutex::new(PublicKeyCache::new(
+                    TESTING_BLS_CACHE_MAX_CAPACITY,
+                ))),
+                network.subscribe_events(),
+            )
+            .await
+        }
+        SyncMode::Full => {
+            SyncerProxy::new_full(
+                blockchain.clone(),
+                Arc::clone(&network),
+                Arc::new(Mutex::new(PublicKeyCache::new(
+                    TESTING_BLS_CACHE_MAX_CAPACITY,
+                ))),
+                zkp_prover.proxy(),
+                network.subscribe_events(),
+            )
+            .await
+        }
+        SyncMode::Light => {
+            SyncerProxy::new_light(
+                blockchain.clone(),
+                Arc::clone(&network),
+                Arc::new(Mutex::new(PublicKeyCache::new(
+                    TESTING_BLS_CACHE_MAX_CAPACITY,
+                ))),
+                zkp_prover.proxy(),
+                network.subscribe_events(),
+            )
+            .await
+        }
+    }
+}
+
+pub async fn sync_two_peers(
+    num_batches_macro_sync: usize,
+    num_batches_live_sync: usize,
+    sync_mode: SyncMode,
+) {
+    let hub = MockHub::default();
+    let mut networks = vec![];
+
+    // Setup first peer.
+    let env1 = VolatileEnvironment::new(11).unwrap();
+    let time = Arc::new(OffsetTime::new());
+    let blockchain1 = Arc::new(RwLock::new(
+        Blockchain::new(
+            env1.clone(),
+            BlockchainConfig::default(),
+            NetworkId::UnitAlbatross,
+            time,
+        )
+        .unwrap(),
+    ));
+
+    // Produce the blocks.
+    let producer = BlockProducer::new(signing_key(), voting_key());
+    produce_macro_blocks_with_txns(&producer, &blockchain1, num_batches_macro_sync, 1, 2);
+
+    let net1: Arc<Network> = TestNetwork::build_network(
+        num_batches_macro_sync as u64 * 10,
+        Default::default(),
+        &mut Some(hub),
+    )
+    .await;
+    networks.push(Arc::clone(&net1));
+    let blockchain1_proxy = BlockchainProxy::from(&blockchain1);
+    let syncer1 = SyncerProxy::new_history(
+        blockchain1_proxy.clone(),
+        Arc::clone(&net1),
+        Arc::new(Mutex::new(PublicKeyCache::new(
+            TESTING_BLS_CACHE_MAX_CAPACITY,
+        ))),
+        net1.subscribe_events(),
+    )
+    .await;
+    let zkp_prover1 = ZKPComponent::new(
+        BlockchainProxy::from(&blockchain1),
+        Arc::clone(&net1),
+        false,
+        Some(zkp_test_exe()),
+        env1.clone(),
+        PathBuf::from(KEYS_PATH),
+    )
+    .await;
+    let consensus1 = Consensus::from_network(
+        env1,
+        BlockchainProxy::from(&blockchain1),
+        Arc::clone(&net1),
+        syncer1,
+        zkp_prover1.proxy(),
+    );
+
+    // Setup second peer (not synced yet).
+    let time = Arc::new(OffsetTime::new());
+    let env2 = VolatileEnvironment::new(11).unwrap();
+
+    let blockchain2_proxy = match sync_mode {
+        SyncMode::History | SyncMode::Full => {
+            let blockchain2 = Arc::new(RwLock::new(
+                Blockchain::new(
+                    env2.clone(),
+                    BlockchainConfig::default(),
+                    NetworkId::UnitAlbatross,
+                    time,
+                )
+                .unwrap(),
+            ));
+            BlockchainProxy::from(blockchain2)
+        }
+        SyncMode::Light => {
+            let blockchain2 = Arc::new(RwLock::new(LightBlockchain::new(NetworkId::UnitAlbatross)));
+            BlockchainProxy::from(blockchain2)
+        }
+    };
+
+    let net2: Arc<Network> = TestNetwork::build_network(
+        num_batches_macro_sync as u64 * 10 + 1,
+        Default::default(),
+        &mut Some(MockHub::default()),
+    )
+    .await;
+    networks.push(Arc::clone(&net2));
+
+    let zkp_prover2 = ZKPComponent::new(
+        blockchain2_proxy.clone(),
+        Arc::clone(&net2),
+        false,
+        Some(zkp_test_exe()),
+        env2.clone(),
+        PathBuf::from(KEYS_PATH),
+    )
+    .await;
+
+    let mut syncer2 = syncer(&sync_mode, &net2, &blockchain2_proxy, &zkp_prover2).await;
+
+    Network::connect_networks(&networks, num_batches_macro_sync as u64 * 10 + 1).await;
+    let zkp_prover2_proxy = zkp_prover2.proxy();
+    spawn(zkp_prover2);
+
+    let macro_sync_result = match syncer2 {
+        SyncerProxy::History(ref mut syncer) => {
+            let macro_sync_result = syncer.macro_sync.next().await;
+            // Now we move the syncing peer to the live sync.
+            syncer.move_peer_into_live_sync(net1.get_local_peer_id());
+            macro_sync_result
+        }
+        SyncerProxy::Light(ref mut syncer) => {
+            let macro_sync_result = syncer.macro_sync.next().await;
+            // Now we move the syncing peer to the live sync.
+            syncer.move_peer_into_live_sync(net1.get_local_peer_id());
+            macro_sync_result
+        }
+        SyncerProxy::Full(ref mut syncer) => {
+            let macro_sync_result = syncer.macro_sync.next().await;
+            // Now we move the syncing peer to the live sync.
+            syncer.move_peer_into_live_sync(net1.get_local_peer_id());
+            macro_sync_result
+        }
+    };
+    log::debug!("Macro sync result {:?}", macro_sync_result);
+    assert_eq!(
+        macro_sync_result,
+        Some(MacroSyncReturn::Good(net1.get_local_peer_id()))
+    );
+
+    let consensus2 = Consensus::new(
+        env2,
+        blockchain2_proxy.clone(),
+        Arc::clone(&net2),
+        syncer2,
+        1,
+        zkp_prover2_proxy,
+    );
+    let consensus2_proxy = consensus2.proxy();
+    let events = blockchain2_proxy.read().notifier_as_stream();
+    let mut events = events.filter(|event| {
+        future::ready(matches!(
+            event,
+            BlockchainEvent::Finalized(_) | BlockchainEvent::EpochFinalized(_)
+        ))
+    });
+    let mut consensus_events = consensus2_proxy.subscribe_events();
+    spawn(consensus2);
+
+    for _ in 0..num_batches_live_sync {
+        produce_macro_blocks_with_txns(&producer, &Arc::clone(&blockchain1), 1, 4, 2);
+        let block = blockchain1_proxy.read().head().clone();
+        _ = net1.publish::<BlockTopic>(block.clone()).await;
+        _ = net1.publish::<BlockHeaderTopic>(block).await;
+        let sync_result = events.next().await;
+        assert!(sync_result.is_some());
+    }
+    let consensus1_proxy = consensus1.proxy();
+    _ = consensus_events.next().await;
+    assert!(consensus2_proxy.is_established());
+    assert_eq!(
+        blockchain2_proxy.read().election_head().block_number(),
+        consensus1_proxy
+            .blockchain
+            .read()
+            .election_head()
+            .block_number(),
+    );
+    assert_eq!(
+        blockchain2_proxy.read().election_head_hash(),
+        consensus1_proxy.blockchain.read().election_head_hash(),
+    );
+    assert_eq!(
+        blockchain2_proxy.read().macro_head_hash(),
+        consensus1_proxy.blockchain.read().macro_head_hash(),
+    );
+
+    match blockchain2_proxy {
+        BlockchainProxy::Full(blockchain) => {
+            assert!(blockchain.read().state.accounts.is_complete(None));
+        }
+        BlockchainProxy::Light(_) => {}
+    }
+}
